Added some robustness to the search for a matlab version in xASL_GUI_Startup.py. Installer script can now never require sudo if the user has all required packages and opts out of a .desktop connection. Thanks & credit to Jan Petr and associated IT personelle for the suggested changes.

+++ xASL_GUI_Linux_Install.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 EXPLOREASLROOT="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 echo "Installing dependencies..."
-sudo apt-get install -y python3.8 libpq-dev python3-dev python3-pip python3-venv python3-wheel pkg-config libfreetype6-dev >/dev/null 2>&1 || exit 1
+PKGLIST=(python3.8 libpq-dev python3-dev python3-pip python3-venv python3-wheel pkg-config libfreetype6-dev)
+for PKG in "${PKGLIST[@]}"; do
+  if [[ $( dpkg-query -l | grep -c "${PKG}" ) -gt 0 ]]; then
+    echo "${PKG} was already installed"
+  else
+    echo "Installing ${PKG}..."
+    sudo apt-get install -y "${PKG}"
+  fi
+done
 
 echo "Setting up virtual environment..."
 mkdir "${EXPLOREASLROOT}/venv" -p
@@ -19,6 +27,18 @@ python3.8 ${EXPLOREASLROOT}/xASL_GUI_run.py
 EOM
 chmod +x "${EXPLOREASLROOT}/xASL_GUI_run.sh"
 
+# Allow user to make the choice of whether they'd like the program to be visible as an application
+read -p "[y|n] Would you like to set up a .desktop program link to make ExploreASL_GUI visible as an application? Note: sudo password will be required. " CHOICE
+while [[ ! "${CHOICE,,}" =~ ^(y|n)$ ]]; do
+  echo "That is not a valid input."
+  read -p "[y|n] Would you like to set up a .desktop program link to make ExploreASL_GUI visible as an application? Note: sudo password will be required. " CHOICE
+done
+if [[ "${CHOICE}" == "n" ]]; then
+  echo "ExploreASL_GUI has been installed."
+  exit 0
+fi
+
+# Assuming the user said yes, make the .desktop file and move it to /usr/share/applications
 echo "Generating .desktop link"
 cat > "${EXPLOREASLROOT}/ExploreASL_GUI.desktop" << EOM
 [Desktop Entry]
